feat: add retry logic and reduce concurrency in analyze-inactivity (#280)
